Amended product data

+++ src/main/java/com/example/ecommerceproject/Entities/Cart.java
@@ -18,7 +18,7 @@ public class Cart {
   @Column(name = "name")
   private String name;
   @Column(name = "price")
-  private int price;
+  private double price;
   @Column(name = "description")
   private String description;
   @Column(name = "image")
@@ -40,11 +40,11 @@ public class Cart {
     this.name = name;
   }
 
-  public int getPrice() {
+  public double getPrice() {
     return price;
   }
 
-  public void setPrice(int price) {
+  public void setPrice(double price) {
     this.price = price;
   }
 
@@ -75,7 +75,7 @@ public class Cart {
   public Cart() {
   }
 
-  public Cart(int count, String name, int price, String description, String image) {
+  public Cart(int count, String name, double price, String description, String image) {
     this.count = count;
     this.name = name;
     this.price = price;
